multi-agent-collaboration-system/requirements.txt
Flask
requests
crewai
langchain-community
duckduckgo-search
websocket-client
